Rename config command to env
